# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asset Management API",
    "summary": "Deterministic portfolio analytics over a currency-aware ledger.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/portfolios": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "List Portfolios",
        "description": "Return every tracked portfolio, for the UI's portfolio picker.\n\nDeclared before ``/{portfolio_id}`` so FastAPI matches this literal path first.",
        "operationId": "list_portfolios_api_v1_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioSummary"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Api V1 Portfolios Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Create Portfolio",
        "description": "Create a new empty portfolio and return its assigned identity.",
        "operationId": "create_portfolio_api_v1_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Portfolio",
        "description": "Return a portfolio's identity + base currency (404 if it doesn't exist).",
        "operationId": "get_portfolio_api_v1_portfolios__portfolio_id__get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/transactions": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "List Transactions",
        "description": "Return the portfolio's full ledger as typed transactions.",
        "operationId": "list_transactions_api_v1_portfolios__portfolio_id__transactions_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction-Output"
                  },
                  "title": "Response List Transactions Api V1 Portfolios  Portfolio Id  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Add Transactions",
        "description": "Append manually-entered ledger events (404 if the portfolio is absent).",
        "operationId": "add_transactions_api_v1_portfolios__portfolio_id__transactions_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Transaction-Input"
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerIngestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/holdings": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "List Holdings",
        "description": "Return the portfolio's tracked securities + classification metadata.",
        "operationId": "list_holdings_api_v1_portfolios__portfolio_id__holdings_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldingInfo"
                  },
                  "title": "Response List Holdings Api V1 Portfolios  Portfolio Id  Holdings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/analytics": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Analytics",
        "description": "Return aggregated cost-basis / realized-P&L / XIRR analytics (404 if absent).",
        "operationId": "get_analytics_api_v1_portfolios__portfolio_id__analytics_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAnalytics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/positions": {
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Add Positions",
        "description": "Record current-holding snapshots as opening BUY lots (404 if absent).\n\nThis is the fast on-ramp for an existing portfolio: assert each position's\nticker, quantity, and cost basis instead of re-keying years of trades.",
        "operationId": "add_positions_api_v1_portfolios__portfolio_id__positions_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PositionSnapshot"
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerIngestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/imports": {
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Create Import",
        "description": "Accept a broker statement and process it in the background.\n\nReturns ``202 Accepted`` immediately with a ``PENDING`` job \u2014 a decade-long\nexport can hold thousands of rows, far too slow to finish inside the request.\nPoll ``GET /portfolios/{id}/imports/{job_id}`` until the status is terminal,\nthen refetch analytics.\n\nThe upload's declared ``content_type`` is untrusted browser metadata, so the\nformat comes from ``source_format`` and the parser validates the actual bytes.\nRe-uploading identical content is rejected with ``409`` (it would double-count\nthe whole ledger); pass ``allow_duplicate=true`` to override deliberately.",
        "operationId": "create_import_api_v1_portfolios__portfolio_id__imports_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "source_format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/StatementFormat",
              "default": "robinhood_csv"
            }
          },
          {
            "name": "allow_duplicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Allow Duplicate"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_import_api_v1_portfolios__portfolio_id__imports_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementImportStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "List Imports",
        "description": "Return the portfolio's import history, newest first.",
        "operationId": "list_imports_api_v1_portfolios__portfolio_id__imports_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatementImportStatus"
                  },
                  "title": "Response List Imports Api V1 Portfolios  Portfolio Id  Imports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}/imports/{job_id}": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Get Import",
        "description": "Return one import job's live status/progress (the endpoint the UI polls).",
        "operationId": "get_import_api_v1_portfolios__portfolio_id__imports__job_id__get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementImportStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_import_api_v1_portfolios__portfolio_id__imports_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_import_api_v1_portfolios__portfolio_id__imports_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HoldingInfo": {
        "properties": {
          "ticker": {
            "type": "string",
            "minLength": 1,
            "title": "Ticker"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          }
        },
        "type": "object",
        "required": [
          "ticker"
        ],
        "title": "HoldingInfo",
        "description": "A tracked security's classification metadata (provider output).\n\nGrouping keys for allocation roll-ups; the market value is joined in later once\nthe market-data layer (\u00a71.3) can price the position."
      },
...
      "LedgerIngestResult": {
        "properties": {
          "portfolio_id": {
            "type": "integer",
            "title": "Portfolio Id"
          },
          "created_transactions": {
            "type": "integer",
            "title": "Created Transactions"
          },
          "tickers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tickers"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "source_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatementFormat"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "created_transactions",
          "tickers"
        ],
        "title": "LedgerIngestResult",
        "description": "What a write/import operation actually recorded (the endpoint's response)."
      },
...
      "PortfolioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "base_currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Base Currency"
          }
        },
        "type": "object",
        "required": [
          "name",
          "base_currency"
        ],
        "title": "PortfolioCreate",
        "description": "Request body to create a new (empty) portfolio."
      },
      "PortfolioSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "base_currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Base Currency"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "base_currency"
        ],
        "title": "PortfolioSummary",
        "description": "Lightweight portfolio identity + reporting base currency (provider output)."
      },
...
      "StatementFormat": {
        "type": "string",
        "enum": [
          "robinhood_csv"
        ],
        "title": "StatementFormat",
        "description": "Supported broker statement layouts for the import endpoint."
      },
...
      "Transaction-Input": {
        "properties": {
          "ticker": {
            "type": "string",
            "minLength": 1,
            "title": "Ticker"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity",
            "default": "0"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Price",
            "default": "0"
          },
          "fees": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Fees",
            "default": "0"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "default": "0"
          },
          "split_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Split Ratio",
            "default": "1"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "type",
          "trade_date",
          "currency"
        ],
        "title": "Transaction",
        "description": "A single dated ledger event for one ticker, in its native currency.\n\nField usage by ``type``:\n\n* ``BUY`` / ``SELL``: ``quantity`` (> 0) and ``price`` per share; ``fees`` optional.\n* ``DIVIDEND``: ``amount`` = total cash received (native currency).\n* ``FEE``: ``amount`` = total fee paid (native currency).\n* ``SPLIT``: ``split_ratio`` (e.g. ``2`` for a 2-for-1); other fields ignored."
      },
      "Transaction-Output": {
        "properties": {
          "ticker": {
            "type": "string",
            "minLength": 1,
            "title": "Ticker"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "default": "0"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price",
            "default": "0"
          },
          "fees": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Fees",
            "default": "0"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "default": "0"
          },
          "split_ratio": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Split Ratio",
            "default": "1"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "type",
          "trade_date",
          "currency"
        ],
        "title": "Transaction",
        "description": "A single dated ledger event for one ticker, in its native currency.\n\nField usage by ``type``:\n\n* ``BUY`` / ``SELL``: ``quantity`` (> 0) and ``price`` per share; ``fees`` optional.\n* ``DIVIDEND``: ``amount`` = total cash received (native currency).\n* ``FEE``: ``amount`` = total fee paid (native currency).\n* ``SPLIT``: ``split_ratio`` (e.g. ``2`` for a 2-for-1); other fields ignored."
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "BUY",
          "SELL",
          "DIVIDEND",
          "FEE",
          "SPLIT"
        ],
        "title": "TransactionType",
        "description": "Ledger event kinds supported by the deterministic calculators."
      },
...
    }
  }
}